.requirements.txt
abnf==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1 ; python_full_version < '3.11'
attrs==25.4.0
authlib==1.6.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
chardet==5.2.0
charset-normalizer==3.4.2
cint==1.0.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==45.0.4
diskcache==5.6.3
docstring-parser==0.16
eval-type-backport==0.3.0
exceptiongroup==1.3.0 ; python_full_version < '3.11'
fastapi==0.115.13
fickling==0.1.5
frozenlist==1.8.0
gitdb==4.0.12
gitpython==3.1.45
google-adk==1.19.0
google-api-core==2.28.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.128.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.34.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.33.0
google-cloud-storage==3.6.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.52.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gql==4.0.0
graphql-core==3.2.7
graphviz==0.21
greenlet==3.2.3 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
intervaltree==3.1.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaitaistruct==0.11
mako==1.3.10
markupsafe==3.0.3
mcp==1.9.4
multidict==6.7.0
networkx==3.4.2 ; python_full_version < '3.11'
networkx==3.6 ; python_full_version >= '3.11'
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.3.0 ; python_full_version >= '3.11'
oauthlib==3.3.1
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pdfminer-six==20250506
pillow==12.0.0
platformdirs==4.5.0
polyfile-weave==0.5.7
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyparsing==3.2.3
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.37.0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
sentry-sdk==2.46.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.41
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
sse-starlette==2.3.6
starlette==0.46.2
stdlib-list==0.11.1
tenacity==9.1.2
tomli==2.2.1 ; python_full_version < '3.11'
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2 ; sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
wandb==0.23.0
watchdog==6.0.0
weave==0.52.17
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
